Add a regressiontest.py script
This can be used to generate a junit xml report when replaying multiple
traces from a directory.
It also prints a nice summary of how many matched.

Currently it cannot handled cases that time out very well (all following
tests will be marked as skipped), but that needs to be fixed in
QuickCheckVEngine rather than in this wrapper script.